@@ -1,12 +1,14 @@
 module github.com/kovey/db-go
 
-go 1.16
+go 1.18
 
 require github.com/go-sql-driver/mysql v1.6.0
 
 require github.com/kovey/debug-go v0.0.2
 
 require github.com/ClickHouse/clickhouse-go v1.4.5
+
+require github.com/cloudflare/golz4 v0.0.0-20150217214814-ef862a3cdc58 // indirect
 
 require (
 	github.com/kovey/config-go v0.0.3
