@@ -7,7 +7,7 @@ func TestDeletePrepare(t *testing.T) {
 	where := NewWhere()
 	where.Like("nickname", "%hello%")
 	del.Where(where)
-	del.WhereByMap(map[string]interface{}{"id": 1, "name": "kovey"}).WhereByList([]string{"last_id > 0"})
+	del.WhereByMap(map[string]any{"id": 1, "name": "kovey"}).WhereByList([]string{"last_id > 0"})
 	t.Logf("sql: %s", del)
 }
 
@@ -16,6 +16,6 @@ func TestCkDeletePrepare(t *testing.T) {
 	where := NewWhere()
 	where.Like("nickname", "%hello%")
 	del.Where(where)
-	del.WhereByMap(map[string]interface{}{"id": 1, "name": "kovey"}).WhereByList([]string{"last_id > 0"})
+	del.WhereByMap(map[string]any{"id": 1, "name": "kovey"}).WhereByList([]string{"last_id > 0"})
 	t.Logf("sql: %s", del)
 }
