@@ -7,38 +7,38 @@ import (
 	"github.com/kovey/db-go/sql"
 )
 
-type TableShardingInterface interface {
-	Database() *sharding.Mysql
-	Insert(interface{}, map[string]interface{}) (int64, error)
-	Update(interface{}, map[string]interface{}, map[string]interface{}) (int64, error)
-	Delete(interface{}, map[string]interface{}) (int64, error)
-	DeleteWhere(interface{}, *sql.Where) (int64, error)
-	BatchInsert(interface{}, []map[string]interface{}) (int64, error)
-	FetchRow(interface{}, map[string]interface{}, interface{}) (interface{}, error)
-	FetchAll(interface{}, map[string]interface{}, interface{}) ([]interface{}, error)
-	FetchAllByWhere(interface{}, *sql.Where, interface{}) ([]interface{}, error)
-	FetchPage(interface{}, map[string]interface{}, interface{}, int, int) ([]interface{}, error)
-	FetchPageByWhere(interface{}, *sql.Where, interface{}, int, int) ([]interface{}, error)
+type TableShardingInterface[T any] interface {
+	Database() *sharding.Mysql[T]
+	Insert(any, map[string]any) (int64, error)
+	Update(any, map[string]any, map[string]any) (int64, error)
+	Delete(any, map[string]any) (int64, error)
+	DeleteWhere(any, sql.WhereInterface) (int64, error)
+	BatchInsert(any, []map[string]any) (int64, error)
+	FetchRow(any, map[string]any, T) (T, error)
+	FetchAll(any, map[string]any, T) ([]T, error)
+	FetchAllByWhere(any, sql.WhereInterface, T) ([]T, error)
+	FetchPage(any, map[string]any, T, int, int) ([]T, error)
+	FetchPageByWhere(any, sql.WhereInterface, T, int, int) ([]T, error)
 }
 
-type TableSharding struct {
+type TableSharding[T any] struct {
 	table string
-	db    *sharding.Mysql
+	db    *sharding.Mysql[T]
 }
 
-func NewTableSharding(table string, isMaster bool) *TableSharding {
-	return &TableSharding{db: sharding.NewMysql(isMaster), table: table}
+func NewTableSharding[T any](table string, isMaster bool) *TableSharding[T] {
+	return &TableSharding[T]{db: sharding.NewMysql[T](isMaster), table: table}
 }
 
-func (t *TableSharding) Database() *sharding.Mysql {
+func (t *TableSharding[T]) Database() *sharding.Mysql[T] {
 	return t.db
 }
 
-func (t *TableSharding) GetTableName(key interface{}) string {
+func (t *TableSharding[T]) GetTableName(key any) string {
 	return fmt.Sprintf("%s_%d", t.table, t.db.GetShardingKey(key))
 }
 
-func (t *TableSharding) Insert(key interface{}, data map[string]interface{}) (int64, error) {
+func (t *TableSharding[T]) Insert(key any, data map[string]any) (int64, error) {
 	in := sql.NewInsert(t.GetTableName(key))
 	for field, value := range data {
 		in.Set(field, value)
@@ -47,7 +47,7 @@ func (t *TableSharding) Insert(key interface{}, data map[string]interface{}) (in
 	return t.db.Insert(key, in)
 }
 
-func (t *TableSharding) Update(key interface{}, data map[string]interface{}, where map[string]interface{}) (int64, error) {
+func (t *TableSharding[T]) Update(key any, data map[string]any, where map[string]any) (int64, error) {
 	up := sql.NewUpdate(t.GetTableName(key))
 	for field, value := range data {
 		up.Set(field, value)
@@ -58,21 +58,21 @@ func (t *TableSharding) Update(key interface{}, data map[string]interface{}, whe
 	return t.db.Update(key, up)
 }
 
-func (t *TableSharding) Delete(key interface{}, where map[string]interface{}) (int64, error) {
+func (t *TableSharding[T]) Delete(key any, where map[string]any) (int64, error) {
 	del := sql.NewDelete(t.GetTableName(key))
 	del.WhereByMap(where)
 
 	return t.db.Delete(key, del)
 }
 
-func (t *TableSharding) DeleteWhere(key interface{}, where *sql.Where) (int64, error) {
+func (t *TableSharding[T]) DeleteWhere(key any, where sql.WhereInterface) (int64, error) {
 	del := sql.NewDelete(t.GetTableName(key))
 	del.Where(where)
 
 	return t.db.Delete(key, del)
 }
 
-func (t *TableSharding) BatchInsert(key interface{}, data []map[string]interface{}) (int64, error) {
+func (t *TableSharding[T]) BatchInsert(key any, data []map[string]any) (int64, error) {
 	batch := sql.NewBatch(t.GetTableName(key))
 	for _, val := range data {
 		in := sql.NewInsert(t.GetTableName(key))
@@ -85,22 +85,22 @@ func (t *TableSharding) BatchInsert(key interface{}, data []map[string]interface
 	return t.db.BatchInsert(key, batch)
 }
 
-func (t *TableSharding) FetchRow(key interface{}, where map[string]interface{}, mt interface{}) (interface{}, error) {
-	return t.db.FetchRow(key, t.GetTableName(key), where, mt)
+func (t *TableSharding[T]) FetchRow(key any, where map[string]any, model T) (T, error) {
+	return t.db.FetchRow(key, t.GetTableName(key), where, model)
 }
 
-func (t *TableSharding) FetchAll(key interface{}, where map[string]interface{}, mt interface{}) ([]interface{}, error) {
-	return t.db.FetchAll(key, t.GetTableName(key), where, mt)
+func (t *TableSharding[T]) FetchAll(key any, where map[string]any, model T) ([]T, error) {
+	return t.db.FetchAll(key, t.GetTableName(key), where, model)
 }
 
-func (t *TableSharding) FetchAllByWhere(key interface{}, where *sql.Where, mt interface{}) ([]interface{}, error) {
-	return t.db.FetchAllByWhere(key, t.GetTableName(key), where, mt)
+func (t *TableSharding[T]) FetchAllByWhere(key any, where sql.WhereInterface, model T) ([]T, error) {
+	return t.db.FetchAllByWhere(key, t.GetTableName(key), where, model)
 }
 
-func (t *TableSharding) FetchPage(key interface{}, where map[string]interface{}, mt interface{}, page, pageSize int) ([]interface{}, error) {
-	return t.db.FetchPage(key, t.GetTableName(key), where, mt, page, pageSize)
+func (t *TableSharding[T]) FetchPage(key any, where map[string]any, model T, page, pageSize int) ([]T, error) {
+	return t.db.FetchPage(key, t.GetTableName(key), where, model, page, pageSize)
 }
 
-func (t *TableSharding) FetchPageByWhere(key interface{}, where *sql.Where, mt interface{}, page, pageSize int) ([]interface{}, error) {
-	return t.db.FetchPageByWhere(key, t.GetTableName(key), where, mt, page, pageSize)
+func (t *TableSharding[T]) FetchPageByWhere(key any, where sql.WhereInterface, model T, page, pageSize int) ([]T, error) {
+	return t.db.FetchPageByWhere(key, t.GetTableName(key), where, model, page, pageSize)
 }
